interactive_charts refactorization
