+++ reporting/templates/comprehensive_report.html
@@ -86,7 +86,6 @@
         <!-- Portfolio Performance -->
         <div class="section">
             <h2>Portfolio Performance</h2>
-            <div id="equity_curve" class="plotly-chart">{{ charts.equity_curve | safe }}</div>
             <div id="metrics_summary">{{ charts.metrics_summary | safe }}</div>
         </div>
